Use webpack, reorganise files per extension component
