+++ .gitignore
@@ -1,4 +1,3 @@
 node_modules
+extension
 web-ext-artifacts
-extension/controller/translation/bergamot-translator-worker.wasm
-extension/controller/translation/bergamot-translator-worker.js
