--- extension/controller/translation/WASMTranslationHelper.js
@@ -1,684 +1,0 @@
-/* eslint-disable no-global-assign */
-/* eslint-disable no-native-reassign */
-/* eslint-disable max-lines */
-
-/**
- * @typedef {Object} TranslationRequest
- * @property {String} from
- * @property {String} to
- * @property {String} text
- * @property {Boolean} html
- * @property {Integer?} priority
- */
-
-const CACHE_NAME = "bergamot-translations";
-
-const MAX_DOWNLOAD_TIME = 60000; // TODO move this
-
-const WASM_TRANSLATION_WORKER_URL = compat.runtime.getURL('controller/translation/WASMTranslationWorker.js');
-
-/**
- * Little wrapper around the message passing API to keep track of messages and
- * their responses in such a way that you can just wait for them by awaiting
- * the promise returned by `request()`.
- */
-class WorkerChannel {
-    constructor(worker) {
-        this.worker = worker;
-        this.worker.onmessage = this.onmessage.bind(this);
-        this.serial = 0;
-        this.pending = new Map();
-    }
-
-    request(message) {
-        return new Promise((resolve, reject) => {
-            const id = ++this.serial;
-            this.pending.set(id, {resolve, reject});
-            this.worker.postMessage({id, message});
-        })
-    }
-
-    onmessage({data: {id, message, error}}) {
-        if (id === undefined)
-            return;
-
-        const {resolve, reject} = this.pending.get(id);
-        this.pending.delete(id);
-
-        if (error !== undefined)
-            reject(error);
-        else
-            resolve(message); // Note: message can be undefined
-    }
-}
-
-/**
- * Wrapper around bergamot-translator and model management. You only need
- * to call translate() which is async, the helper will manage execution by
- * itself.
- */
- class WASMTranslationHelper {
-    
-    /**
-     * @param {{
-     *  cacheSize: Number?,
-     *  useNativeIntGemm: Boolean?,
-     *  workers: Number?,
-     *  batchSize: Number?
-     * }} options
-     */
-    constructor(options) {
-        this.options = options || {};
-
-        // registry of all available models and their urls: Promise<List<Model>>
-        this.registry = lazy(async () => {
-            try {
-                return await this.loadModelRegistery();
-            } catch (error) {
-                throw new Error(`Could not download model registry: ${error.message}`);
-            }
-        });
-
-        /**
-         * Map of downloaded model data files as buffers per model.
-         * @type {Map<{from:String,to:String}, Promise<Map<String,ArrayBuffer>>>}
-         */
-        this.buffers = new Map();
-        
-        /**
-         * A map of language-pairs to a list of models you need for it.
-         * @type {Map<{from:String,to:String}, Promise<{from:String,to:String}[]>>}
-         */
-        this.models = new Map();
-
-        /**
-         * @type {Array<{idle:Boolean, worker:Proxy}>} List of active workers
-         * (and a flag to mark them idle or not)
-         */
-        this.workers = [];
-
-        /**
-         * Maximum number of workers
-         * @type {Number} 
-         */
-        this.workerLimit = Math.max(this.options.workers || 0, 1);
-
-        /**
-         * List of batches we push() to & shift() from using `enqueue`.
-         * @type {{
-         *    id: Number,
-         *    key: String,
-         *    priority: Number,
-         *    models: TranslationModel[],
-         *    requests: Array<{
-         *      request: TranslationRequest,
-         *      resolve: (response: TranslationResponse),
-         *      reject: (error: Error)
-         *    }>
-         * }}
-         */
-        this.queue = [];
-
-        /**
-         * batch serial to help keep track of batches when debugging
-         * @type {Number}
-         */
-        this.batchSerial = 0;
-
-        /**
-         * Number of requests in a batch before it is ready to be translated in
-         * a single call. Bigger is better for throughput (better matrix packing)
-         * but worse for latency since you'll have to wait for the entire batch
-         * to be translated.
-         * @type {Number}
-         */
-        this.batchSize = Math.max(this.options.batchSize || 8, 1);
-
-        /**
-         * Error handler for all errors that are async, not tied to a specific
-         * call and that are unrecoverable.
-         * @type {(error: Error)}
-         */
-        this.onerror = err => console.error('WASM Translation Worker error:', err);
-    }
-
-    /**
-     * Loads a worker thread, and wraps it in a message passing proxy. I.e. it
-     * exposes the entire interface of TranslationWorker here, and all calls
-     * to it are async. Do note that you can only pass arguments that survive
-     * being copied into a message. Returns Proxy<TranslationWorker>.
-     */
-    loadWorker() {
-        // TODO is this really not async? Can I just send messages to it from
-        // the start and will they be queued or something?
-        const worker = new Worker(WASM_TRANSLATION_WORKER_URL);
-        worker.onerror = this.onerror.bind(this);
-
-        // Initialisation options
-        worker.postMessage({options: this.options});
-
-        // Little wrapper around the message passing api of Worker to make it
-        // easy to await a response to a sent message.
-        const channel = new WorkerChannel(worker);
-
-        // Wrap the worker in a Proxy so you can treat it as if it is an
-        // instance of the TranslationWorker class that lives inside the worker.
-        // All function calls to it are transparently passed through the message
-        // passing channel.
-        return new Proxy(worker, {
-            get(target, name, receiver) {
-                return (...args) => {
-                    return channel.request({name, args: Array.from(args)}) // returns a Promise
-                }
-            }
-        });
-    }
-
-    /**
-     * Loads the model registry. Uses the registry shipped with this extension,
-     * but formatted a bit easier to use, and future-proofed to be swapped out
-     * with a TranslateLocally type registry. Returns Promise<List<Model>>
-     */
-    async loadModelRegistery() {
-        const cache = typeof caches !== 'undefined' ? caches.open(CACHE_NAME) : null;
-        const response = await fetch('https://translatelocally.com/models.json');
-        const {models} = await response.json();
-        let serial = 0;
-
-        // Check whether each of the models is already downloaded
-        models.forEach((model) => {
-            // Give the model an id for easy look-up
-            model.id = ++serial;
-            
-            // Async check whether the model is in cache
-            let modelInCache = false;
-            if (cache) {
-                cache.then(cache => cache.match(model.url).then(response => {
-                    modelInCache = response && response.ok;
-                }));
-            }
-
-            Object.defineProperty(model, 'local', {
-                enumerable: true,
-                get: () => modelInCache || this.hasTranslationModel(model)
-            });
-        });
-
-        // Add 'from' and 'to' keys for each model. Since theoretically a model
-        // can have multiple froms keys in TranslateLocally, we do a little
-        // product here.
-        return Array.from(flatten(models, function*(model) {
-            try {
-                const to = first(Intl.getCanonicalLocales(model.trgTag));
-                for (let from of Intl.getCanonicalLocales(Object.keys(model.srcTags))) {
-                    yield {from, to, model};
-                }
-            } catch (err) {
-                console.log('Skipped model', model, err);
-            }
-        }));
-    }
-
-    hasTranslationModel({from, to}) {
-        const key = JSON.stringify({from, to});
-        return this.buffers.has(key);
-    }
-
-    /**
-     * Gets or loads translation model data. Caching wrapper around
-     * `loadTranslationModel()`. Returns the same promise type.
-     */
-    getTranslationModel({from, to}) {
-        const key = JSON.stringify({from, to});
-
-        if (!this.buffers.has(key))
-            this.buffers.set(key, this.loadTranslationModel({from, to}));
-
-        return this.buffers.get(key);
-    }
-
-    downloadModel(id) {
-        return new PromiseWithProgress(async (accept, reject, update) => {
-            try {
-                const model = (await this.registry).find(({model}) => model.id === id);
-                accept(await this.loadTranslationModel(model, update));
-            } catch (err) {
-                reject(err);
-            }
-        });
-    }
-
-    /**
-     * Downloads (or from cache) a translation model and returns a set of
-     * ArrayBuffers. These can then be passed to a TranslationWorker thread
-     * to instantiate a TranslationModel inside the WASM vm.
-     * Returns Promise<Map<str,ArrayBuffer>>.
-     */
-    async loadTranslationModel({from, to}, update) {
-        performance.mark(`loadTranslationModule.${JSON.stringify({from, to})}`);
-
-        // Find that model in the registry which will tell us about its files
-        const entries = (await this.registry).filter(model => model.from == from && model.to == to);
-
-        // Prefer tiny models above non-tiny ones (right now base models don't even work properly 😅)
-        entries.sort(({model: a}, {model: b}) => (a.shortName.indexOf('tiny') === -1 ? 1 : 0) - (b.shortName.indexOf('tiny') === -1 ? 1 : 0));
-
-        if (!entries)
-            throw new Error(`No model for '${from}' -> '${to}'`);
-
-        const entry = first(entries).model;
-
-        const compressedArchive = await this.getItemFromCacheOrWeb(entry.url, entry.checksum, update);
-
-        const archive = pako.inflate(compressedArchive);
-
-        const files = await untar(archive.buffer);
-
-        const find = (filename) => {
-            const found = files.find(file => file.name.match(/(?:^|\/)([^\/]+)$/)[1] === filename)
-            if (found === undefined)
-                throw new Error(`Could not find '${filename}' in model archive`);
-            return found;
-        };
-
-        const config = YAML.parse(find('config.intgemm8bitalpha.yml').readAsString());
-
-        const model = find(config.models[0]).buffer;
-
-        const vocabs = config.vocabs.map(vocab => find(vocab).buffer);
-
-        const shortlist = find(config.shortlist[0]).buffer;
-
-        performance.measure('loadTranslationModel', `loadTranslationModule.${JSON.stringify({from, to})}`);
-
-        // Return the buffers
-        return {model, vocabs, shortlist, config};
-    }
-
-    /**
-     * Helper to either get a URL remote or from cache. Downloaded file is
-     * always checked against checksum.
-     * @param {String} url
-     * @param {String} checksum sha256 checksum as hexadecimal string
-     * @returns {Promise<ArrayBuffer>}
-     */
-    async getItemFromCacheOrWeb(url, checksum, update) {
-        try {
-            let cache = null;
-
-            // Only check the cache if we're not running in private mode
-            if (typeof caches !== 'undefined') {
-                cache = await caches.open(CACHE_NAME);
-                const match = await cache.match(url);
-
-                if (match) {
-                    // Found it in the cache, let's check whether it (still) matches the
-                    // checksum. If not, redownload it.
-                    const buffer = await match.arrayBuffer();
-                    if (await this.digestSha256AsHex(buffer) === checksum)
-                        return buffer;
-                    else
-                        cache.delete(url);
-                }
-            }
-
-            return await this.getItemFromWeb(url, checksum, cache, update);
-        } catch (e) {
-            throw new Error(`Failed to download '${url}': ${e.message}`);
-        }
-    }
-
-    /**
-     * Helper to download file from the web (and store it in the cache if that
-     * is passed in as well). Verifies the checksum.
-     * @param {String} url
-     * @param {String} checksum sha256 checksum as hexadecimal string
-     * @param {Cache?} cache optional cache to save response into
-     * @returns {Promise<ArrayBuffer>}
-     */
-    async getItemFromWeb(url, checksum, cache, update) {
-        try {
-            // Rig up a timeout cancel signal for our fetch
-            const abort = new AbortController();
-            let timeout = setTimeout(() => abort.abort(), MAX_DOWNLOAD_TIME);
-
-            // Start downloading the url, will give us response headers.
-            const response = await fetch(url, {signal: abort.signal});
-
-            // Also stream it to cache. We check the promise later, but we have
-            // to clone the response before we start reading.
-            const cached = cache ? cache.put(url, response.clone()) : Promise.resolve(true);
-
-            /** @type {ArrayBuffer} */
-            let buffer;
-
-            if (update) {
-                const size = parseInt(response.headers.get('Content-Length'));
-
-                // If we didn't get a size, we can't report progress.
-                if (size) {
-                    const body = new Uint8Array(size);
-                    const reader = response.body.getReader();
-                    let read = 0;
-
-                    while (true) {
-                        const {done, value} = await reader.read();
-
-                        if (done)
-                            break;
-
-                        body.set(value, read);
-                        read += value.length;
-                        update({size, read});
-
-                        // Reset timeout because progress is good!
-                        clearTimeout(timeout);
-                        timeout = setTimeout(() => abort.abort(), MAX_DOWNLOAD_TIME);
-                    }
-
-                    buffer = body.buffer;
-                }
-            }
-
-            // Fallback fast route when we either don't have an update() or we
-            // didn't get a Content-Length header.
-            if (!buffer)
-                buffer = await response.arrayBuffer();
-
-            // Download finished, remove the abort timer
-            clearTimeout(timeout);
-
-            // Make sure caching succeeded (if no cache, always true)
-            await cached;
-
-            // Checking checksum afterwards. Previously I used sub-resource
-            // integrity to check the hash, but that will delay getReader()
-            // until the resource has been downloaded & checked.
-            if (await this.digestSha256AsHex(buffer) !== checksum)
-                throw new TypeError(`Response for ${url} did not match checksum ${checksum}`);
-            
-            return buffer;
-        } catch (e) {
-            // If the download timed out or didn't pass muster, make sure it
-            // doesn't end up in the cache in a bad way.
-            if (cache)
-                cache.delete(url);
-            throw e;
-        }
-    }
-
-    /**
-     * Expects ArrayBuffer, returns String.
-     * @param {ArrayBuffer} buffer
-     * @returns {Promise<String>} SHA256 checksum as hexadecimal string
-     */
-    async digestSha256AsHex(buffer) {
-        // hash the message
-        const hashBuffer = await crypto.subtle.digest("SHA-256", buffer);
-        // convert buffer to byte array
-        const hashArray = Array.from(new Uint8Array(hashBuffer));
-        // convert bytes to hex string
-        return hashArray.map(b => b.toString(16).padStart(2, "0")).join("");
-    }
-
-    /**
-     * Crappy named method that gives you a list of models to translate from
-     * one language into the other. Generally this will be the same as you
-     * just put in if there is a direct model, but it could return a list of
-     * two models if you need to pivot through a third language.
-     * Returns just [{from:str,to:str}...]. To be used something like this:
-     * ```
-     * const models = await this.getModels(from, to);
-     * models.forEach(({from, to}) => {
-     *   const buffers = await this.loadTranslationModel({from,to});
-     *   [TranslationWorker].loadTranslationModel({from,to}, buffers)
-     * });
-     * ```
-     * @param {String} from
-     * @param {String} to
-     * @returns {Promise<TranslationModel[]>}
-     */
-    getModels(from, to) {
-        const key = JSON.stringify({from, to});
-
-        // Note that the `this.models` map stores Promises. This so that
-        // multiple calls to `getModels` that ask for the same model will
-        // return the same promise, and the actual lookup is only done once.
-        // The lookup is async because we need to await `this.registry`
-        if (!this.models.has(key))
-            this.models.set(key, this.findModels(from, to));
-
-        return this.models.get(key);
-    }
-
-    /**
-     * Find model (or model pair) to translate from `from` to `to`.
-     * @param {String} from
-     * @param {String} to
-     * @returns {Promise<TranslationModel[]>}
-     */
-    async findModels(from, to) {
-        const registry = await this.registry;
-
-        // TODO: This all scales really badly.
-
-        let direct = [], outbound = [], inbound = [];
-
-        registry.forEach(model => {
-            if (model.from === from && model.to === to)
-                direct.push(model);
-            else if (model.from === from)
-                outbound.push(model);
-            else if (model.to === to)
-                inbound.push(model);
-        });
-
-        if (direct.length)
-            return [direct[0]];
-
-        // Find the pivot language
-        const shared = intersect(
-            outbound.map(model => model.to),
-            inbound.map(model => model.from)
-        );
-
-        if (!shared.size)
-            throw new Error(`No model available to translate from '${from}' to '${to}'`);
-
-        return [
-            outbound.find(model => shared.has(model.to)),
-            inbound.find(model => shared.has(model.from))
-        ];
-    }
-
-    /**
-     * Makes sure queued work gets send to a worker. Will delay it till `idle`
-     * to make sure the batches have been filled to some degree. Will keep
-     * calling itself as long as there is work in the queue, but it does not
-     * hurt to call it multiple times. This function always returns immediately.
-     */
-    notify() {
-        requestIdleCallback(async () => {
-            // Is there work to be done?
-            if (!this.queue.length)
-                return;
-
-            // Find an idle worker
-            let worker = this.workers.find(worker => worker.idle);
-
-            // No worker free, but space for more?
-            if (!worker && this.workers.length < this.workerLimit) {
-                try {
-                    worker = {
-                        idle: true,
-                        worker: this.loadWorker()
-                    };
-                    this.workers.push(worker);
-                } catch (e) {
-                    this.onerror(new Error(`Could not initialise translation worker: ${e.message}`));
-                }
-            }
-
-            // If no worker, that's the end of it.
-            if (!worker)
-                return;
-
-            // Up to this point, this function has not used await, so no
-            // chance that another call stole our batch since we did the check
-            // at the beginning of this function and JavaScript is only
-            // cooperatively parallel.
-            const batch = this.queue.shift();
-
-            // Put this worker to work, marking as busy
-            worker.idle = false;
-            try {
-                await this.consumeBatch(batch, worker.worker);
-            } catch (e) {
-                batch.requests.forEach(({reject}) => reject(e));
-            }
-            worker.idle = true;
-
-            // Is there more work to be done? Do another idleRequest
-            if (this.queue.length)
-                this.notify();
-        }, {timeout: 1000}); // Start after 1000ms even if not idle
-    }
-
-    /**
-     * The only real public call you need!
-     * ```
-     * const {translation:str} = await this.translate({
-     *   from: 'de',
-     *   to: 'en',
-     *   text: 'Hallo Welt!',
-     *   html: false, // optional
-     *   priority: 0 // optional, like `nice` lower numbers are translated first
-     * })
-     * ```
-     * @param {TranslationRequest} request
-     * @returns {Promise<TranslationResponse>}
-     */
-    translate(request) {
-        const {from, to, html, priority} = request;
-
-        return new Promise(async (resolve, reject) => {
-            try {
-                // Batching key: only requests with the same key can be batched
-                // together. Think same translation model, same options.
-                const key = JSON.stringify({from, to});
-
-                // (Fetching models first because if we would do it between looking
-                // for a batch and making a new one, we end up with a race condition.)
-                const models = await this.getModels(from, to);
-                
-                // Put the request and its callbacks into a fitting batch
-                this.enqueue({key, models, request, resolve, reject, priority});
-
-                // Tell a worker to pick up the work at some point.
-                this.notify();
-            } catch (e) {
-                reject(e);
-            }
-        });
-    }
-
-    /**
-     * Prune pending requests by testing each one of them to whether they're
-     * still relevant. Used to prune translation requests from tabs that got
-     * closed.
-     * @param {(request:TranslationRequest) => boolean} filter evaluates to true if request should be removed
-     */
-    remove(filter) {
-        const queue = this.queue;
-
-        this.queue = [];
-
-        queue.forEach(batch => {
-            batch.requests.forEach(({request, resolve, reject}) => {
-                if (filter(request)) {
-                    // Add error.request property to match response.request for
-                    // a resolve() callback. Pretty useful if you don't want to
-                    // do all kinds of Funcion.bind() dances.
-                    reject(Object.assign(new Error('removed by filter'), {request}));
-                    return;
-                }
-
-                this.enqueue({
-                    key: batch.key,
-                    priority: batch.priority,
-                    models: batch.models,
-                    request,
-                    resolve,
-                    reject
-                });
-            });
-        });
-    }
-
-    /**
-     * Internal function used to put a request in a batch that still has space.
-     * Also responsible for keeping the batches in order of priority. Called by
-     * `translate()` but also used when filtering pending requests.
-     * @param {{request:TranslateRequest, models:TranslationModel[], key:String, priority:Number?, resolve:(TranslateResponse)=>any, reject:(Error)=>any}}
-     */
-    enqueue({key, models, request, resolve, reject, priority}) {
-        if (priority === undefined)
-            priority = 0;
-         // Find a batch in the queue that we can add to
-         // (TODO: can we search backwards? that would speed things up)
-        let batch = this.queue.find(batch => {
-            return batch.key === key
-                && batch.priority === priority
-                && batch.requests.length < this.batchSize
-        });
-
-        // No batch or full batch? Queue up a new one
-        if (!batch) {
-            batch = {id: ++this.batchSerial, key, priority, models, requests: []};
-            this.queue.push(batch);
-            this.queue.sort((a, b) => a.priority - b.priority);
-        }
-
-        batch.requests.push({request, resolve, reject});
-    }
-
-    /**
-     * Internal method that uses a worker thread to process a batch. You can
-     * wait for the batch to be done by awaiting this call. You should only
-     * then reuse the worker otherwise you'll just clog up its message queue.
-     */
-    async consumeBatch(batch, worker) {
-        performance.mark('BTConsumeBatch.start');
-
-        // Make sure the worker has all necessary models loaded. If not, tell it
-        // first to load them.
-        await Promise.all(batch.models.map(async ({from, to}) => {
-            if (!await worker.hasTranslationModel({from, to})) {
-                const buffers = await this.getTranslationModel({from, to});
-                await worker.loadTranslationModel({from, to}, buffers);
-            }
-        }));
-
-        // Call the worker to translate. Only sending the actually necessary
-        // parts of the batch to avoid trying to send things that don't survive
-        // the message passing API between this thread and the worker thread.
-        const responses = await worker.translate({
-            models: batch.models.map(({from, to}) => ({from, to})),
-            texts: batch.requests.map(({request: {text, html}}) => ({text, html}))
-        });
-
-        // Responses are in! Connect them back to their requests and call their
-        // callbacks.
-        batch.requests.forEach(({request, resolve, reject}, i) => {
-            // TODO: look at response.ok and reject() if it is false
-            resolve({
-                request, // Include request for easy reference? Will allow you
-                         // to specify custom properties and use that to link
-                         // request & response back to each other.
-                ...responses[i] // {target: {text: String}}
-            });
-        });
-        
-        performance.measure('BTConsumeBatch', 'BTConsumeBatch.start');
-    }
-}
