+++ src/benchmark/benchmark.html
@@ -103,15 +103,6 @@
 		<p>Time (excl. start-up) is the total time it takes to translate the test-set, given that the translation model is already loaded and primed.</p>
 		<p>WPS is word count (HTML tags are stripped, excluded elements are stripped as well) divided by the Time column</p>
 
-		<script src="../../controller/translation/compat.js"></script>
-		<script src="../../3rd_party/pako/pako_inflate.js"></script>
-		<script src="../../3rd_party/js-untar/untar.js"></script>
-		<script src="../../controller/translation/func.js"></script>
-		<script src="../../controller/translation/promise.js"></script>
-		<script src="../../controller/translation/yaml.js"></script>
-		<script src="../../controller/translation/TLTranslationHelper.js"></script>
-		<script src="../../controller/translation/WASMTranslationHelper.js"></script>
-		<script src="common.js"></script>
 		<script src="benchmark.js"></script>
 	</body>
 </html>
