+++ src/options/options.html
@@ -46,8 +46,6 @@
         <p>Mainly useful for benchmarking and debugging. <a data-bind:href="benchmarkURL" target="blank_">Benchmark page</a></p>
       </div>
     </div>
-    <script src="../../controller/translation/compat.js"></script>
-    <script src="common.js"></script>
     <script src="options.js"></script>
   </body>
 </html>
