+++ src/popup/popup.html
@@ -94,8 +94,6 @@
 			Record translation requests
 			<a href="#" id="export-recorded-pages-btn" data-bind:hidden="!canExportPages">Export <span data-bind:text-content="recordedPagesCount"></span> pages</button>
 		</label>
-		<script src="../../controller/translation/compat.js"></script>
-		<script src="common.js"></script>
 		<script src="popup.js"></script>
 	</body>
 </html>
