+++ src/shared/func.js
@@ -1,7 +1,7 @@
 /**
  * Little wrapper to delay a promise to be made only once it is first awaited on
  */
-function lazy(factory) {
+export function lazy(factory) {
     let promise = null;
 
     return {
@@ -24,7 +24,7 @@ function lazy(factory) {
  * Array.prototype.map, but with a twist: the functor returns an iterator
  * (or more usefully) a generator, it will then add each of those elements.
  */
-function *flatten(iterable, functor) {
+export function *flatten(iterable, functor) {
     for (let item of iterable)
         yield* functor(item);
 }
@@ -32,7 +32,7 @@ function *flatten(iterable, functor) {
 /**
  * `product([a,b], [1,2]) == [[a,1], [a,2], [b,1], [b,2]]`
  */
-function* product(as, bs) {
+export function* product(as, bs) {
     for (let a of as)
         for (let b of bs)
             yield [a, b];
@@ -42,7 +42,7 @@ function* product(as, bs) {
  * Take the first element from anything that can be iterated over. Like arr[0]
  * or iterable[Symbol.iterator].next().value. If the iterator is empty, throw.
  */
-function first(iterable) {
+export function first(iterable) {
     for (let item of iterable)
         return item;
     throw new RangeError('Iterable is empty');
@@ -51,7 +51,7 @@ function first(iterable) {
 /**
  * Returns a set that is the intersection of two iterables
  */
-function intersect(a, b) {
+export function intersect(a, b) {
     const bSet = new Set(b);
     return new Set(Array.from(a).filter(item => bSet.has(item)));
 }
