+++ src/shared/promise.js
@@ -2,7 +2,7 @@
  * Promise class, but with a progress notification. Useful for things like
  * donwloads where there is information about progress.
  */
-class PromiseWithProgress extends Promise {
+export class PromiseWithProgress extends Promise {
     #listeners;
 
     constructor(factory) {
